REFACTOR-#0000: move 'read_callback' from dispatchers into parsers
